End of section 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -53,11 +54,12 @@ namespace AuctionService.Controllers
         return _mapper.Map<AuctionDto>(auction);
        }
 
+     [Authorize]
      [HttpPost]
      public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto){
 
        var auction = _mapper.Map<Auction>(auctionDto);
-       auction.Seller = "test";
+       auction.Seller = User.Identity.Name;
 
        _context.Auctions.Add(auction);
        
@@ -72,10 +74,15 @@ namespace AuctionService.Controllers
        return CreatedAtAction(nameof(GetAuctionById),new {auction.Id},_mapper.Map<AuctionDto>(auction));
      }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id,UpdateAuctionDto updateAuctionDto){
+
        var auction = await _context.Auctions.Include(x=>x.Item).FirstOrDefaultAsync(x=>x.Id==id);
+
        if(auction == null) return NotFound();
+
+       if(auction.Seller != User.Identity.Name) return Forbid();
 
        auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
        auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -90,6 +97,7 @@ namespace AuctionService.Controllers
        return BadRequest("Problem saving changes");
     }
 
+     [Authorize]
      [HttpDelete("{id}")]
      public async Task<ActionResult> DeleteAuction(Guid id){
 
@@ -97,9 +105,11 @@ namespace AuctionService.Controllers
 
       if(auction == null) return NotFound();
 
+      if(auction.Seller != User.Identity.Name) return Forbid();
+
       _context.Auctions.Remove(auction);
 
-       await _publishEndpoint.Publish(_mapper.Map<AuctionDeleted>(new {Id=auction.Id.ToString()}));
+       await _publishEndpoint.Publish<AuctionDeleted>(new {Id=auction.Id.ToString()});
 
       var result = await _context.SaveChangesAsync() > 0;
